% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advRlab04-package.R
\docType{package}
\name{advRlab04-package}
\alias{advRlab04}
\alias{advRlab04-package}
\title{advRlab04: Advanced R Lab 4}
\description{
This package contains an implementation of a Linear Regression class for the purpose of fulfilling Lab 4 in Advanced R class.
}
\keyword{internal}
