\name{advRlab04-package}
\alias{advRlab04-package}
\alias{advRlab04}
\docType{package}
\title{
\packageTitle{advRlab04}
}
\description{
\packageDescription{advRlab04}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{advRlab04}
\packageIndices{advRlab04}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{advRlab04}

Maintainer: \packageMaintainer{advRlab04}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
