% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simgehugo.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Model with Linear Algebra}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{a linear regression model}

\item{data}{a dataframe}
}
\value{
returns a RC class
}
\description{
This function do the linear regression
}
\examples{
linreg(Petal.Length~Species, data = iris)
}
